Reviews by Product cleanup (#773)
