+++ assets/js/blocks/reviews-by-product/utils.js
@@ -20,10 +20,6 @@ export function renderReview( attributes, review = {}, i = 0 ) {
 	const showProductRating = Number.isFinite( rating ) && showProductRatingAttr;
 
 	const classes = classNames( 'wc-block-reviews-by-product__item', {
-		'has-avatar': showAvatar,
-		'has-date': showReviewDate,
-		'has-name': showReviewerName,
-		'has-rating': showProductRating,
 		'is-loading': isLoading,
 	} );
 	const starStyle = {
